Another float8 to prevent query-crash

+++ internal/signaling/stores/postgres.go
@@ -464,7 +464,7 @@ func (s *PostgresStore) UpdatePeerGeo(ctx context.Context, peerID string, lat, l
 		UPDATE peers
 		SET
 			geo = CASE
-				WHEN $1 IS NOT NULL AND $2 IS NOT NULL THEN ll_to_earth($1::float8, $2::float8)
+				WHEN $1::float8 IS NOT NULL AND $2::float8 IS NOT NULL THEN ll_to_earth($1::float8, $2::float8)
 				ELSE NULL
 			END,
 			updated_at = $3
